more size_t usage instead of int

+++ src/Analyser/Controller.h
@@ -38,7 +38,7 @@ public:
     bool InsertIntoCorpus ( const std::string& inputPath, const std::string& outputPath, const bool newReplacesExisting );
 
 private:
-    std::vector<int> MergeDatasets ( Utilities::DataSet& newDataset, const Utilities::DataSet& existingDataset, const bool newReplacesExisting );
+    std::vector<size_t> MergeDatasets ( Utilities::DataSet& newDataset, const Utilities::DataSet& existingDataset, const bool newReplacesExisting );
 
     bool SearchDirectory ( const std::string& directory, std::vector<std::string>& files );
